requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
appdirs==1.4.4
argh==0.26.2
asgiref==3.2.10
asn1crypto==1.3.0
astroid==2.3.3
astropy==4.0
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.3.0
audioread==2.1.8
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
blinker==1.4
bokeh==1.4.0
boto==2.49.0
boto3==1.14.23
botocore==1.17.23
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
Click==7.0
clldutils==3.5.2
cloudpickle==1.3.0
clyent==1.2.2
cmake==3.18.2.post1
colorama==0.4.3
colorlog==4.2.1
conda==4.8.4
conda-build==3.18.11
conda-package-handling==1.7.0
conda-verify==3.4.2
configparser==5.0.0
contextlib2==0.6.0.post1
coremltools==3.4
cryptography==2.8
csvw==1.7.0
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.11.0
datalab==1.2.0
dateparser==0.7.6
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.2
distlib==0.3.0
distributed==2.11.0
distro==1.5.0
Django==3.1.1
dlib==19.20.0
docutils==0.16
entrypoints==0.3
enum34==1.1.10
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastcache==1.1.0
ffpyplayer==4.3.1
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
fsspec==0.6.2
future==0.18.2
gast==0.3.3
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-cloud-monitoring==0.31.1
google-pasta==0.2.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.29.0
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
httplib2==0.18.1
hypothesis==5.5.4
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.4.2
imagesize==1.2.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446408836/work
imutils==0.5.3
inflect==4.1.0
intervaltree==3.0.2
invoice2data==0.3.5
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
jeepney==0.4.2
Jinja2==2.11.1
jmespath==0.10.0
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
jupyterthemes==0.20.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.1.0
kiwisolver==1.1.0
kraken==2.0.8
lazy-object-proxy==1.4.3
lesscpy==0.14.0
libarchive-c==2.8
librosa==0.8.0
lief==0.9.0
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
more-itertools==8.2.0
moviepy==1.0.0
mpmath==1.1.0
mPyPl==0.0.3.7
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233707986/work
numpydoc==0.9.2
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.4.0.42
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.1
pafy==0.5.5
pandas==1.0.1
pandas-profiling==1.4.0
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pbr==5.4.5
pdftotext==2.1.4
pep8==1.7.1
pexpect==4.8.0
phonemizer==2.2.1
pickleshare==0.7.5
Pillow==7.0.0
pipe==1.5.0
pkginfo==1.5.0.1
playsound==1.2.2
plotly==4.8.2
pluggy==0.13.1
ply==3.11
pooch==1.1.1
proglog==0.1.9
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.12.3
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==4.0.1
pydub==0.24.1
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyQt5==5.15.0
PyQt5-sip==12.8.1
pyrsistent==0.15.7
PySocks==1.7.1
pytesseract==0.3.4
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-bidi==0.4.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
python-utils==2.4.0
python-vlc==3.0.11115
python3-weather-api==0.0.3rc0
pyttsx3==2.90
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
regex @ file:///tmp/build/80754af9/regex_1593435542598/work
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rfc3986==1.4.0
rope==0.16.0
rsa==4.0
Rtree==0.9.3
ruamel-yaml==0.15.87
s3transfer==0.3.3
sacremoses==0.0.43
scikit-build==0.11.1
scikit-image==0.16.2
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1598376882706/work
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
segments==2.1.3
Send2Trash==1.5.0
sentencepiece==0.1.91
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
SoundFile==0.10.3.post1
soupsieve==1.9.5
SpeechRecognition==3.8.1
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
sqlparse==0.3.1
statsmodels==0.11.0
stevedore==2.0.0
sympy==1.5.1
tables==3.6.1
tabulate==0.8.7
tblib==1.6.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-datasets==3.1.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-hub==0.8.0
tensorflow-metadata==0.22.2
tensorflow-utils==0.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-utils==1.0.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tokenizers==0.8.1rc1
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
toolz==0.10.0
torch==1.5.1
torchvision==0.6.1
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
transformers==3.0.2
typed-ast==1.4.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1592847887441/work
tzlocal==2.1
ujson==1.35
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
utils==1.0.1
virtualenv==20.0.21
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
watchdog==0.10.2
watermark==2.0.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
youtube-dl==2020.7.28
zict==1.0.0
zipp==2.2.0
zope.event==4.4
zope.interface==4.7.1
